Rename time tests
