+++ tests/Time/DiffTest.php
@@ -8,39 +8,38 @@ use DigitalCraftsman\DateTimeParts\Time;
 use PHPUnit\Framework\TestCase;
 
 /** @coversDefaultClass \DigitalCraftsman\DateTimeParts\Time */
-final class TimeIsMidnightTest extends TestCase
+final class DiffTest extends TestCase
 {
     /**
      * @test
      *
      * @dataProvider dataProvider
      *
-     * @covers ::isMidnight
+     * @covers ::diff
      */
-    public function is_midnight_works(
-        bool $expectedResult,
+    public function diff_works(
+        \DateInterval $expectedResult,
         Time $time,
+        Time $timeToDiff,
     ): void {
         // -- Act & Assert
-        self::assertSame($expectedResult, $time->isMidnight());
+        self::assertEquals($expectedResult, $time->diff($timeToDiff));
     }
 
     /**
      * @return array<string, array{
-     *   0: boolean,
+     *   0: \DateInterval,
      *   1: Time,
+     *   2: Time,
      * }>
      */
     public function dataProvider(): array
     {
         return [
-            'at midnight' => [
-                true,
-                Time::fromString('00:00:00'),
-            ],
-            'not at midnight' => [
-                false,
+            '5 minutes' => [
+                (new \DateTimeImmutable('2022-10-08 15:00:00'))->diff(new \DateTimeImmutable('2022-10-08 15:05:00')),
                 Time::fromString('15:00:00'),
+                Time::fromString('15:05:00'),
             ],
         ];
     }
