+++ tests/Time/ToStringTest.php
@@ -1,0 +1,26 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigitalCraftsman\DateTimeParts\Time;
+
+use DigitalCraftsman\DateTimeParts\Time;
+use PHPUnit\Framework\TestCase;
+
+/** @coversDefaultClass \DigitalCraftsman\DateTimeParts\Time */
+final class ToStringTest extends TestCase
+{
+    /**
+     * @test
+     *
+     * @covers ::__toString
+     */
+    public function to_string_works(): void
+    {
+        // -- Arrange & Act
+        $time = Time::fromString('15:00:00');
+
+        // -- Assert
+        self::assertEquals('15:00:00.000000', (string) $time);
+    }
+}
